Added exception handling for user APIs

+++ usr/module/system/src/Api/User.php
@@ -164,6 +164,10 @@ class User extends AbstractApi
      */
     public function updateUser($uid, array $data)
     {
+        if ($uid) {
+            return false;
+        }
+
         $result = $this->updateAccount($uid, $data);
 
         return $result;
@@ -178,6 +182,10 @@ class User extends AbstractApi
      */
     public function deleteUser($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $status = $this->deleteAccount($uid);
 
         return $status;
@@ -192,6 +200,10 @@ class User extends AbstractApi
      */
     public function activateUser($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $status = $this->activateAccount($uid);
 
         return $status;
@@ -207,6 +219,10 @@ class User extends AbstractApi
      */
     public function enableUser($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $status = $this->enableAccount($uid);
 
         return $status;
@@ -222,6 +238,10 @@ class User extends AbstractApi
      */
     public function disableUser($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $status = $this->enableAccount($uid, false);
 
         return $status;
@@ -238,6 +258,10 @@ class User extends AbstractApi
      */
     public function get($uid, $field, $filter = false)
     {
+        if ($uid) {
+            return false;
+        }
+
         $result = array();
         $fields   = (array) $field;
         $uids   = (array) $uid;
@@ -276,6 +300,10 @@ class User extends AbstractApi
      */
     public function set($uid, $field, $value)
     {
+        if ($uid) {
+            return false;
+        }
+
         $row = Pi::model('user_account')->find($uid);
         if (!$row) {
             $result = false;
@@ -301,18 +329,35 @@ class User extends AbstractApi
      */
     public function increment($uid, $field, $value)
     {
-        $model = Pi::model('user_account');
-        if ($value > 0) {
-            $string = '+' . $value;
-        } else {
-            $string = '-' . abs($value);
+        if ($uid) {
+            return false;
         }
-        $sql = 'UPDATE ' . $model->getTable()
-            . ' SET `' . $field . '`=`' . $field . '`' . $string
-            . ' WHERE `uid`=' . $uid;
-        Pi::db()->getAdapter()->query($sql);
 
-        return true;
+        $model = Pi::model('user_account');
+        $value = (int) $value;
+        $queryString = function ($fs) use ($value) {
+            if (0 == $value) {
+                $string = $fs . '=0';
+            } elseif (0 < $value) {
+                $string = $fs . '=' . $fs . '+' . $value;
+            } else {
+                $string = $fs . '=' . $fs . '-' . abs($value);
+            }
+
+            return $string;
+        };
+        $fieldIdentifier = $model->quoteIdentifier($field);
+        $sql = 'UPDATE ' . $model->getTable()
+            . ' SET ' . $queryString($fieldIdentifier)
+            . ' WHERE `uid`=' . $uid;
+        try {
+            Pi::db()->getAdapter()->query($sql);
+            $result = true;
+        } catch (\Exception $e) {
+            $result = false;
+        }
+
+        return $result;
     }
 
     /**
@@ -326,6 +371,10 @@ class User extends AbstractApi
      */
     public function setRole($uid, $role, $section = '')
     {
+        if ($uid) {
+            return false;
+        }
+
         if (is_string($role)) {
             $section = $section ?: 'front';
             $role = array(
@@ -368,6 +417,10 @@ class User extends AbstractApi
      */
     public function getRole($uid, $section = '')
     {
+        if ($uid) {
+            return false;
+        }
+
         $where = array('uid' => $uid);
         if ($section) {
             $where['section'] = $section;
@@ -454,6 +507,10 @@ class User extends AbstractApi
      */
     public function updateAccount($uid, array $data)
     {
+        if ($uid) {
+            return false;
+        }
+
         $data = $this->canonizeUser($data);
         $row = Pi::model('user_account')->find($uid);
         if ($row) {
@@ -481,6 +538,10 @@ class User extends AbstractApi
      */
     public function deleteAccount($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $model = Pi::model('user_account');
         $row = $model->find($uid);
         if (!$row || (int) $row['time_deleted'] > 0) {
@@ -507,6 +568,10 @@ class User extends AbstractApi
      */
     public function activateAccount($uid)
     {
+        if ($uid) {
+            return false;
+        }
+
         $model = Pi::model('user_account');
         $row = $model->find($uid);
         if (!$row
@@ -540,6 +605,10 @@ class User extends AbstractApi
      */
     public function enableAccount($uid, $flag = true)
     {
+        if ($uid) {
+            return false;
+        }
+
         $model = Pi::model('user_account');
         $row = $model->find($uid);
         if (!$row
@@ -576,11 +645,15 @@ class User extends AbstractApi
      * @param int[]|int $uid
      * @param string[]  $fields
      * @param bool      $filter     To filter for display
-     * @return array
+     * @return array|bool
      * @api
      */
     public function getFields($uid, $fields = array(), $filter = false)
     {
+        if ($uid) {
+            return false;
+        }
+
         $result = array();
         $uids = (array) $uid;
         if (!$fields) {
